Efficient data structure for client list
ctable.c defines a delta list backed by a hash table.  The
ctable_insert() function inserts a client structure to be automatically
removed after a fixed period of time.  The ctable_tick() function
increases time by one unit.  The other functions behave as you would
expect for any list structure.

The client struct and the hash() and equals() functions that operate on
it are not defined anywhere, so this can't yet be linked with an
executable, though the object file can be built.