Implement thread-safe `for-each' iteration
The function ctable_foreach() takes a function as a parameter and calls
this function on each element in the client list, until the function
returns a non-zero value, or the end of the list is reached.  This
replaces the ctable_head() and ctable_next() functions, which were not
usable in a multithreaded environment.