Overhaul of the response end of the protocol
Responses now consist of a JSON object header and (potentially) a body.
The header contains one mandatory field and potentially many optional
fields.  The mandatory field is "status", which may be either "okay" or
"error".  If the message has a body, the "size" field--an integer
indicating the number of bytes to read after the header--is also
required.

The CONNECT and DISCONNECT commands return the bare minimum response (a
header with a status field only).  The LIST command returns a header
with the size field filled in, and a message body which consists of a
JSON array representing the client list, as before.

Though the new protocol makes heavier use of JSON, the cJSON library is
no longer used.  Constructing JSON objects is easy enough.