Implement DISCOVER directory command
DISCOVER is like LIST except that a port number is specified so that the
entry corresponding to the requesting client's IP address and the
specified port number can be ignored.  This is useful since the router
probably doesn't want to know about its own entry in the table.  The
LIST command remains as a convenience for clients.