"search" method becomes "broadcast", & more
The search method is renamed "broadcast" since the router doesn't care
about the payload--it just broadcasts it throughout the network.

udp_send_msg() is moved from nodeservice.c to udpserver.c/udp.h, so that
it can be used in other modules (in particular, it replaces the
send_message() function in router.c).

Additionally, the process_broadcast() function now sets the "hop-port"
field in broadcast messages.  To accomodate this in a zero-copy way,
clients should send a blank (all space characters) string of length 5 as
the value of the hop-port field.