Use void* in delta list, add delta_update()
Delta lists now store void pointers so that they can be used to store
any pointer type.  A delta_update() function is added which behaves
identically to the old delta_insert() function.  The semantics of
delta_insert() are modified so that delta_insert() never increases the
delta of an element that is already present in the list (this is what
delta_update() is for).