Cache broadcast msg ids to prevent duplicates
Broadcast messages are expected to contain an "id" field which should
be a random (as random as possible) identifier for that particular
message.  Routers remember this field and refuse to forward subsequent
messages with the same id for some period of time after it initially
sees the id.

Additionally, some fixes for the valgrind targets in both makefiles.