Set up TCP server for router
The router now accepts the "info" method over TCP.  Currently it
responds with no useful information.  In the future it might return the
current load, any protocol extensions it supports, etc.

Code in both nodeservice.c and dirservice.c has also been improved in a
number of minor ways.  The implementation of error responses is now
shared between the directory and the router, and annoying
inconsistencies in the connection/message handlers have been smoothed
out (to some extent).