Use variadic function to get JSON values
The function jsmn_get_values takes any number of char*-int* pairs,
setting the values of the int*s to the index of the JSON value
associated with the corresponding char* (i.e. the key).  The variadic
arguments should be terminated with a null pointer.

The function returns -1 if any of the keys were not found, or 0 if all
of the keys were found.  If the function returns -1, then the int*s
associated with the keys which were not found are set to -1.