Restrict use of response_node, process IP requests
Use of the response_node struct and related functions is restricted to
dealing with variable-length JSON output.  In cases where the number of
nodes is known before hand, it is better to construct the response on
the stack and send it using generic functions (sprintf and
tcp_send_bytes).

A new "ip" request method is added to the router.  The router will
return the requesting client's IP address in response.