Several improvements to output
The double CRLF sequence is appended consistently to responses.  The
"info" method (router) now returns the number of clients connected and
the number of messages in the message cache.  This required adding a
"size" member to the delta_list structure, in order to keep track of the
number of elements in the list.