Don't require "hop-port" field in bcast messages
The "hop-port" field is no longer required or used by the router.  This
field was a blemish on an otherwise clean protocol and served little
purpose.  Short forwarding loops are prevented by refusing to forward a
message to the same IP address it came from, regardless of the port.

This is somewhat problematic if two routers are running on the same host
(e.g. for testing).  A solution is to simply disable the check--the
message caching function should prevent short loops anyways.